Tell Ruby openssl to trust server certificate
This makes it un-necessary to put the certificate file in the OpenSSL
system directory and execute `c_rehash` in order to avoid
"OpenSSL::SSL::SSLError: SSL_read:: sslv3 alert bad record mac" errors
showing up in the logs when running locally. The error was generally
associated with failing to load JS or CSS files which can be
disconcerting.

This solution came from StackOverflow at
http://stackoverflow.com/questions/14748701/how-to-set-ca-bundle-path-for-openssl-in-ruby