Create `onStart` for each transloadit form
The `assembly` data passed into the old `uploadStart` function did not
always have `fields` data set on it and as a result the form could not
be found based on the idx property. In order to avoid this situation the
onStart callback is now a closure created with a reference to the form
in scope. The calls to start the spinner was changed to always create a
new spin object because `spin.start()` was behaving unreliably.

Fixes #134