Refactor `AdminAccount`; validation in `before_POST`
Oh, and handle errors because they could happen and they weren't being
handled at all. So errors are now displayed to the user and the
submitted values are retained (since they aren't saved unless there are
no errors). With this refactor the validation logic is done prior to the
work in the `#before_POST` method, the business logic is contained in
`#post_account`.