Use transmission type instead of name
The expectation is customers will be more interested in the
transmission type than the exact details of it. Additionally DataOne
does not provide guarentees about the content of transmission name while
there are only three options for the transmission type. This allows the
transmission type to be predictably used for translation and
presentation.

See #131