Add a task to start the application with stock WEBrick
This enables local testing of the application with a longer lived
process. Shotgun is great for development because it keeps reloading the
application when ruby files change but to do this each request is
handled by a new process which means objects which need to be longer
lived in order to be properly tested (specifically thinking about
Librat/Metrics here) can not properly be exercised.

See #99