Log uncaught exceptions in `Bunny::Channel`
See #138 - https://sprint.ly/product/16794/item/138

Exceptions thrown in the consumer (for example, due to missing classes)
were not being exposed to STDOUT because the `Bunny::Channel` lived in a
child process. This was making some bugs difficult to diagnose.
Introducing the `#process_uncaught_exception` method and giving it as a
`Proc` to `Bunny::Channel#on_uncaught_exception` allows the
`Lynr::Worker` to log these previously swallowed errors.