Change how document is prepared to send to Elasticsearch
See #138 - https://sprint.ly/product/16794/item/138

The elasticsearch index has no need of the image or raw VIN import data,
so don't send it along. Additionally the dealership_id should be indexed
as a string not as an object id or whatever it was before. dealership id
was represented in the index as `Lynr::Model::Vehicle.dealership.$oid`
and it is now `Lynr::Model::Vehicle.dealership`.