Create controller and view for searching and results
See #139 - https://sprint.ly/product/16794/item/139
See #140 - https://sprint.ly/product/16794/item/140

Searches are performed as GET requests as they are about retrieving data
and not sending it for storage. The `Admin::Search#search` method is
responsible for executing the search based on the terms in
`req.params['q']`. The search performed is a filtered search where
vehciles (documents) are filtered by the dealership value and a missing
deleted_at date. Search results are displayed like they would be on the
manage vehicles screens with the addition of a page title / header above
the results stating what was searched and how many results were found.