Add some helper methods to make specs easier
See #203 - https://sprint.ly/product/16794/item/203

Generating valid requests to run through controllers is a pain in the
ass and the new `#request` helper method makes it a bit easier without
having to set everything to use the RouteHelper. Specifically this makes
it so the `Sly::Route` doesn't need to be created if we aren't testing
via `Sly::Route#call`.