Use a pagination data hash in Inventory
See #209 - https://sprint.ly/product/16794/item/209

Rather than creating separate variables for each piece of data the
pagination component needs create a single `Hash` which can be passed
render_partial. This could be moved to `Lynr::Controller::Paginated` but
it would have to take a lot of information as parameters anyway so at
this point the abstraction may not be worth it.