Add a shitty no-purpose element to fix a visual bug
Fixes #201 - https://sprint.ly/product/16794/item/201

When the viewport is wider than 700px the header of the `.main` element
doesn't extend to the edges. By adding this `b.hat.hat-window` element
we can create the illusion of having the header extend to the edges
without having to re-work the CSS and page structure for this "edge
case".