Move `Demo` module into `Signup` controller
See #197 - https://sprint.ly/product/16794/item/197

The `Demo` module it not useful without the context of the `Signup`
controller so move the module into `Signup`. Additionally, since it is
not yet possible (because Ruby 1.9.3) to override a class method with a
module method, move the default `Signup` controller logic into a
`Default` module within Signup. The modules are used to dynamically
extend the `Signup` controller when it is `initialize`d based on the
demo feature flag.