Significant rework of `Lynr::Metrics`
See #197 - https://sprint.ly/product/16794/item/197

Many methods in `Lynr::Metrics` accepted a config object in order to
test how the class reacted when some of the librato configuration was
missing. This was necessary before the ConfigHelper shared_context. This
refactoring still checks for partial configuration but handles it
better (by using a dummy queue when not enabled). Additionally, the
specs are now more consistent with the usage pattern in the application
rather than being artificial because of the needs of testing.