`#method_missing` no longer errors if property isn't configured
See #197 - https://sprint.ly/product/16794/item/197

Using the default (`super`) method_missing behavior when a property
wasn't defined in the configuration was too frustrating. In most cases
partial configuration is tested for or handled gracefully so crashing
the application because a property wasn't defined (especially common in
testing) was ridiculous.