Create `Default` and `Demo` modules for `AdminBilling` controller
See #197 - https://sprint.ly/product/16794/item/197

Though the billing link should be hidden on the demo site, if a user
edits the URL it is possible to find it. If the billing screen is found
display some text instead of the billing form which wouldn't be
processed successfully even if a card were entered (because a Stripe
customer doesn't exist).