Create method to convert live domain to demo domain
See #269 - https://sprint.ly/product/16794/item/269

In order to not require separate queue processing infrastructure for
each instance of the Lynr web application it is necessary to translate
from the live domain to the demo domain when events are of the demo
type. This feels wrong but I don't see a better way to handle the
situation without either:

1. having separate queue processors with config matching application
   configuration
2. passing the whole application configuration as part of the event

The first solution is the 'right' way but would require expensive,
mostly idle, heroku dynos for each demo site (idle particularly for the
staging site) or a new deployment infrastructure such that the
application (web or queue processor) could run on
self-hosted/self-administered hardware. And while I think going down the
path of self-hosted will happen I'm not prepared to commit to the
infrastructure and planning necessary to accomplish it yet. This
statement is particularly ironic when considering I spent two days
working on a 'better' solution to the problem, which presumably would
have been enough time to get a solid start on a different self-hosted
infrastructure setup.