Don't include color information from xml feed
See #247 - https://sprint.ly/product/16794/item/247

Because decoding a VIN does not give us a specific set of color
information this data was being presented as a comma separated list of
possible values. A better solution would be to create a combo-box with
data from the VIN decoding but in the interest of time it has simply
been removed.