Don't require journal acknowledgment
This re-assessment was prompted by failing specs on CircleCI. The specs
failed because journaling on CircleCI containers wasn't enabled.
However, the default write concern is that one node has acknowledged the
write so the data is in the replica set. Requiring journal
acknowledgment was also slowing down database writes because it was
waiting 1/3 of the `journalCommitInterval` configuration setting (see
http://docs.mongodb.org/v2.4/reference/configuration-options/#journalCommitInterval).