Handle uncaught `StandardError` in `Worker::Job#perform`
`Worker::Job#perform` wraps the execution of `Queue::Job#perform` for
each job passed to `Worker::Job#process`. This allows `StandardError`s
raised by `Queue::Job#perform` to be rescued and wrapped in a failing
`JobResult` allowing the `Worker` to continue despite errors.

This moves the rescue logic out of individual `Queue::Job` sub-classes
and handles the situation in a general way.